Load API key from pickle file
- Created function to load API key from a pickle file.

Reason:
API key isn't public, so we can't expose it in a public repository like this.
The solution was to store it in a file, more specifically in a pickle file type.
Whenever we run the imdb_movie.py, it loads the key from this pickle file by calling the function load_api_key() and then stores the returned value in a variable, such as key.