Added next/previous movie feature
After showing the first movie info, the program asks to the user if he/she wants to move to the next movie.
The user navigates cyclically, meaning that if the user is checkint the latest movie and select 'next', the list goes to the beginnig.
The same for the 'prev' case.
If user types anything different from above cases, the program exits.